Add sensor simulations for Eddiebot base
This commit adds sensor simulations for the Eddiebot base in the robot description. The included sensor macros for ultrasonic and IR sensors are utilized to define the positions and orientations of the sensors relative to the base link. The sensors are named and assigned specific parent links, topics, and joint names. This enables accurate simulation of ultrasonic and IR sensors on the Eddiebot base.

Changes:

- Added ultrasonic sensor simulations for left and right sensors on the Eddiebot base
- Added IR sensor simulations for front, left, and right sensors on the Eddiebot base
- Configured sensor positions and orientations relative to the base link for accurate simulation